# api/openapi.yaml
openapi: 3.1.0
info:
  title: One API
  version: 0.2.0
paths:
...
  /users/me:
    get:
      tags:
      - users
      summary: Get Me
      operationId: get_me_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - users
      summary: Patch Me
      operationId: patch_me_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /todos:
    get:
      tags:
      - todos
      summary: List Todos
      operationId: list_todos_todos_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TodoResponse'
                type: array
                title: Response List Todos Todos Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - todos
      summary: Create Todo
      operationId: create_todo_todos_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /todos/{todo_id}:
    patch:
      tags:
      - todos
      summary: Patch Todo
      operationId: patch_todo_todos__todo_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: todo_id
        in: path
        required: true
        schema:
          type: string
          title: Todo Id
      - name: x-client-updated-at
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Updated-At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - todos
      summary: Delete Todo
      operationId: delete_todo_todos__todo_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: todo_id
        in: path
        required: true
        schema:
          type: string
          title: Todo Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /reflections:
    get:
      tags:
      - reflections
      summary: List Reflections
      operationId: list_reflections_reflections_get
      security:
      - HTTPBearer: []
      parameters:
      - name: period_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Period Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReflectionNoteResponse'
                title: Response List Reflections Reflections Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - reflections
      summary: Create Reflection
      operationId: create_reflection_reflections_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReflectionWriteRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReflectionNoteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reflections/{reflection_id}:
    delete:
      tags:
      - reflections
      summary: Delete Reflection
      operationId: delete_reflection_reflections__reflection_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: reflection_id
        in: path
        required: true
        schema:
          type: string
          title: Reflection Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /today:
    get:
      tags:
      - today
      summary: Get Today
      operationId: get_today_today_get
      security:
      - HTTPBearer: []
      parameters:
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodayResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /today/order:
    put:
      tags:
      - today
      summary: Put Today Order
      operationId: put_today_order_today_order_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodayOrderWriteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodayResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReflectionNoteResponse:
      properties:
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        period_type:
          type: string
          title: Period Type
        period_start:
          type: string
          format: date
          title: Period Start
        period_end:
          type: string
          format: date
          title: Period End
        content:
          type: string
          title: Content
        sentiment:
          type: string
          title: Sentiment
        tags:
          items:
            type: string
          type: array
          title: Tags
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - period_type
      - period_start
      - period_end
      - content
      - sentiment
      - tags
      - created_at
      - updated_at
      title: ReflectionNoteResponse
    ReflectionWriteRequest:
      properties:
        period_type:
          type: string
          title: Period Type
        period_start:
          type: string
          format: date
          title: Period Start
        period_end:
          type: string
          format: date
          title: Period End
        content:
          type: string
          title: Content
        sentiment:
          type: string
          title: Sentiment
          default: okay
        tags:
          items:
            type: string
          type: array
          title: Tags
      type: object
      required:
      - period_type
      - period_start
      - period_end
      - content
      title: ReflectionWriteRequest
    TodayItemResponse:
      properties:
        item_type:
          type: string
          title: Item Type
        item_id:
          type: string
          title: Item Id
        title:
          type: string
          title: Title
        category_id:
          type: string
          title: Category Id
        completed:
          type: boolean
          title: Completed
        sort_bucket:
          type: integer
          title: Sort Bucket
        sort_score:
          type: number
          title: Sort Score
        subtitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle
        is_pinned:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Pinned
        priority:
          anyOf:
          - type: integer
          - type: 'null'
          title: Priority
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        preferred_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: Preferred Time
      type: object
      required:
      - item_type
      - item_id
      - title
      - category_id
      - completed
      - sort_bucket
      - sort_score
      title: TodayItemResponse
    TodayOrderItemRequest:
      properties:
        item_type:
          type: string
          title: Item Type
        item_id:
          type: string
          title: Item Id
        order_index:
          type: integer
          title: Order Index
      type: object
      required:
      - item_type
      - item_id
      - order_index
      title: TodayOrderItemRequest
    TodayOrderWriteRequest:
      properties:
        date_local:
          type: string
          format: date
          title: Date Local
        items:
          items:
            $ref: '#/components/schemas/TodayOrderItemRequest'
          type: array
          title: Items
      type: object
      required:
      - date_local
      - items
      title: TodayOrderWriteRequest
    TodayResponse:
      properties:
        date_local:
          type: string
          format: date
          title: Date Local
        items:
          items:
            $ref: '#/components/schemas/TodayItemResponse'
          type: array
          title: Items
        completed_count:
          type: integer
          title: Completed Count
        total_count:
          type: integer
          title: Total Count
        completion_ratio:
          type: number
          title: Completion Ratio
      type: object
      required:
      - date_local
      - items
      - completed_count
      - total_count
      - completion_ratio
      title: TodayResponse
    TodoCreateRequest:
      properties:
        category_id:
          type: string
          title: Category Id
        title:
          type: string
          title: Title
        notes:
          type: string
          title: Notes
          default: ''
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        priority:
          type: integer
          title: Priority
          default: 50
        is_pinned:
          type: boolean
          title: Is Pinned
          default: false
      type: object
      required:
      - category_id
      - title
      title: TodoCreateRequest
    TodoResponse:
      properties:
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        category_id:
          type: string
          title: Category Id
        title:
          type: string
          title: Title
        notes:
          type: string
          title: Notes
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        priority:
          type: integer
          title: Priority
        is_pinned:
          type: boolean
          title: Is Pinned
        status:
          type: string
          title: Status
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - category_id
      - title
      - notes
      - due_at
      - priority
      - is_pinned
      - status
      - completed_at
      - created_at
      - updated_at
      title: TodoResponse
    TodoUpdateRequest:
      properties:
        category_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Category Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        priority:
          anyOf:
          - type: integer
          - type: 'null'
          title: Priority
        is_pinned:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Pinned
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: TodoUpdateRequest
...
    UserResponse:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          format: email
          title: Email
        apple_sub:
          anyOf:
          - type: string
          - type: 'null'
          title: Apple Sub
        display_name:
          type: string
          title: Display Name
        timezone:
          type: string
          title: Timezone
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - email
      - display_name
      - timezone
      - created_at
      title: UserResponse
    UserUpdateRequest:
      properties:
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
      type: object
      title: UserUpdateRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer